And the same for udivdi3, [TPG]: you are a ning-nong

+++ Usermode/Libraries/libgcc.so_src/libgcc.c
@@ -53,7 +53,7 @@ uint64_t __udivdi3(uint64_t Num, uint64_t Den)
 	if(Den == 0)	__asm__ __volatile__ ("int $0x0");
 	// Common speedups
 	if(Num <= 0xFFFFFFFF && Den <= 0xFFFFFFFF)
-		return Num / Den;
+		return (uint32_t)Num / (uint32_t)Den;
 	if(Den == 1)	return Num;
 	if(Den == 2)	return Num >> 1;
 	if(Den == 16)	return Num >> 4;
